+++ CHANGES.md
@@ -1,3 +1,10 @@
+= 0.0.2
+
+* 修正 DMG 创建脚本未正确使用 App 名称的问题；
+* 配置新增 `project` 项，用于直接指定项目文件（XX.xcworkspace 或 XX.xcodeproj）；
+* 修正 Cocoaoutfile 无法初始化的问题；
+* 新增 cocoaout build 命令，用于仅编译项目；
+
 = 0.0.1 
 
 * 实现基本的编译，打包 DMG 功能